verifying tests, some cleanup
