+++ src/weather/Weather.tsx
@@ -1,26 +1,23 @@
 import { CityWeather } from "../models/CityWeather";
-import React from "react";
+import { tileTemperatureColor } from "./tileTemperatureColor";
 import "./weather.css";
 
-
+const removeWeather = (cityWeather:CityWeather) => {
+  const items = JSON.parse(localStorage.getItem("favoriteItems") || "[]");
+  const filteredItems = items.filter((item: any) => item.name !== cityWeather?.name);
+  localStorage.setItem(
+    "favoriteItems",
+    JSON.stringify(filteredItems, null, 2)
+  );
+}
 
 const Weather = ({ cityWeather }: { cityWeather: CityWeather | undefined }) => {
-  const tileColor = (temperature: string) => {
-    const temp = parseInt(temperature);
-    if (temp > 30) {
-      return "is-danger";
-    } else if (temp > 20) {
-      return "is-warning";
-    } else {
-      return "is-info";
-    }
-  }
-
   if (cityWeather) {
     return (
       <div className="column is-one-quarter">
-        <div className={"is-block city weather-container tile notification " + tileColor(cityWeather.temperature) + " has-text-centered"}>
-          <h3 className="subtitle mb-2">{cityWeather.name}</h3>
+        <div className={"is-block city weather-container tile notification " + tileTemperatureColor(cityWeather.temperature) + " has-text-centered"}>
+          <button className="remove" onClick={() => removeWeather(cityWeather)}>Remove</button>
+          <h3 className="name subtitle mb-2">{cityWeather.name}</h3>
           <div className="details">
             <span className="temperature title">{cityWeather.temperature}</span>
             <div className="weather mt-2">
