+++ src/weather/weather.css
@@ -1,3 +1,10 @@
 .weather-container {
   height: 100%;
 }
+
+.remove {
+  position: absolute;
+  top: 0;
+  right: 0;
+  cursor: pointer;
+}
